Fix native sendmessage signature and call for static analyzer warnings

+++ src/Shared/NativeMethods.cs
@@ -11,6 +11,6 @@ namespace SyncroSim.Epi
         public const int EM_SETCUEBANNER = 0x1501;
 
         [DllImport("user32.dll", CharSet = CharSet.Auto)]
-        public static extern int SendMessage(IntPtr hWnd, int msg, int wParam, [MarshalAs(UnmanagedType.LPWStr)]string lParam);
+        public static extern IntPtr SendMessage(IntPtr hWnd, int msg, IntPtr wParam, [MarshalAs(UnmanagedType.LPWStr)]string lParam);
     }
 }
